examples/langgraph-chatanthropic-agent/pyproject.toml
[project]
name = "langgraph-chatanthropic-agent"
version = "0.1.0"
description = "LangGraph ChatAnthropic Agent with Helicone integration"
requires-python = ">=3.9"
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
bypass-selection = true
include = ["main.py"] 